Update README.md
Advanced analytics is not just one particular model of technology, it is a collection of various powerful techniques used to gain actionable insights for better decision-making. In the evolving world of business, data has become the center of decision-making for any organization. Companies nowadays utilize their vast pool of information to extract meaningful insights with the help of advanced analytics rather than relying on their intuition. It is a robust tool, surpassing the limitations of conventional methods with its enhanced capabilities, utilizing modern techniques to uncover patterns, trends and correlations within the data. 
READ MORE- https://www.iconresources.com/unleashing-the-power-of-advanced-analytics-transforming-data-into-strategic-insights/